Check if user owns the supplement

+++ apps/api/src/app/domain/supplement/supplement.controller.ts
@@ -1,21 +1,52 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, UseGuards } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpException,
+  HttpStatus,
+  Param,
+  Post,
+  Put,
+  UseGuards,
+} from '@nestjs/common';
 import { Supplement } from './supplement.schema';
 import { CreateSupplementDto } from './dto/createSupplement.dto';
 import { UpdateSupplementDto } from './dto/updateSupplement.dto';
 import { SupplementService } from './supplement.service';
 import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
+import { UserService } from '../user/user.service';
 @Controller('supplement')
 export class SupplementController {
-  constructor(private readonly supplementService: SupplementService) {}
+  constructor(
+    private readonly supplementService: SupplementService,
+    private readonly userService: UserService
+  ) {}
 
+  @UseGuards(JwtAuthGuard)
   @Post()
-  async create(@Body() supplementDto: CreateSupplementDto): Promise<Supplement> {
+  async create(
+    @Body() supplementDto: CreateSupplementDto
+  ): Promise<Supplement> {
     return this.supplementService.create(supplementDto);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Put(':id')
-  async edit(@Param('id') id: string, supplementDto: UpdateSupplementDto): Promise<Supplement> {
-    return this.supplementService.edit(id, supplementDto);
+  async edit(
+    @Param('id') params,
+    supplementDto: UpdateSupplementDto
+  ): Promise<Supplement> {
+    if (
+      (await this.userService.getCurrentId()) !==
+      (await this.supplementService.findOne(params.id)).createdById
+    ) {
+      throw new HttpException(
+        'Can only edit owned supplements',
+        HttpStatus.FORBIDDEN
+      );
+    }
+    return this.supplementService.edit(params.id, supplementDto);
   }
 
   @Get()
@@ -31,9 +62,15 @@ export class SupplementController {
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
   async remove(@Param() params): Promise<Supplement> {
+    if (
+      (await this.userService.getCurrentId()) !==
+      (await this.supplementService.findOne(params.id)).createdById
+    ) {
+      throw new HttpException(
+        'Can only delete owned supplements',
+        HttpStatus.FORBIDDEN
+      );
+    }
     return this.supplementService.remove(params.id);
   }
-
 }
-
-
